Add files via upload
This is specifically for re-practicing, so as to find out which code lines, and comment lines can be removed. 

Also to ensure that the screenshot images can go to the Selenium_Screenshot_Images for the new project, by manually creating the Java files, then copying and pasting the required code lines from the original files (and to avoid copying and pasting the original Java files into the new project's folders).

Thank you.